tokio: upgrade to tokio v1.42.0
This lets us get rid of the local modifications to Tokio. That said, a
patch needs to be added to mio for the tests to compile. A fix will be
sent upstream.

Also update tokio-macros and mio in lockstep.

Tokio depends on nix 0.29, but we only have 0.28. This is okay since
the 0.29 changes only matter on FreeBSD. Nix is not upgraded here since
that would require crosvm changes.

Test: m
Change-Id: Ibf18a174c5bd98d9bf0ee4acf75dcea23b2fba49
Signed-off-by: Alice Ryhl <aliceryhl@google.com>