Add missing rules.mk files.
For the diplomat and diplomat_core crates.

This was probably the result of not running "git add" after changing
cargo_embargo.json and running `crate_tool regenerate`. Unfortunately,
the AyeAye analyzer isn't able to detect missing files like this. The
pre-upload check could detect this, but currently fails to. This is a
bug.

Test: ./crate_tool regenerate diplomat diplomat_core
Change-Id: I1df96ee5e671ee650262da38b6fc6c027bb457fe