Remove Cargo.lock files that aren't present in the vendored code.
I'm not sure how these were introduced. My theory is a Rust analyzer
running on the crate directories that had a race condition with
`./crate_tool regenerate` and managed to create the Cargo.lock after
moving the new crate contents from the temporary directory in out/ and
before regenerating the checksum file. The solution is probably to do
all work in the temporary directory.

Test: ./crate_tool regenerate serde unicode-bidi
Change-Id: I51e3824e13a5e4b6dc5cea1403e816a678661dbf