Add a memory safe alternative to libz
This crate can be used directly from C and C++.

This change includes a patch including two commits from upstream,
squashed here because they cannot be cleanly cherry-picked because
they span two subcrates, one of which is not included in Android.
The first commit includes the header files in a subcrate not included
in Android, the second commit moves them into the libz-rs-sys crate
which we use in Android.

This includes:
https://github.com/trifectatechfoundation/zlib-rs/pull/385
 - include `zlib.h` and `zconf.h`
 - move .h files to the `libz-rs-sys` crate

Bug: 420420630
Test: atest
Test: Move some targets from libz to libz_rs_ffi, build and run tests
locally.

Change-Id: Ifd72e4a460c640466d990ba4765b6c2b6e926aeb