aarch64-paging: Disable hwasan
This crate intentionally synthesizes virtual addresses from physical
addresses. This means that those reconstituted pointers will not have a
tag byte attached. It is intended for bare metal usage, where either
hwasan is not actually available, or where it would be implemented *in
terms of* this crate, so enabling it is not appropriate.

Bug: 376927508
Test: TH
Change-Id: I418d5ca9bdba031c28c29d794bd4c78f48b579e4