Work around long 32-bit x86 compile for libregex_automata
This build is taking ~25 minutes to compile locally for 32-bit x86, so
we just reduce the optimization level to 0 to prevent the optimizer from
spinning unnecessarily. This is intended as a temporary workaround,
rather than a permanent fix, but it makes things better for local
development while the real bug can be fixed.

Bug: http://b/436672173
Test: m libregex_automata
Change-Id: I45212e61d82b8c183a7042fc38d13798b534b041